requirements.txt
Django~=5.2
pip~=25.0.1
attrs~=25.3.0
PySocks~=1.7.1
filelock~=3.18.0
asgiref~=3.8.1
sqlparse~=0.5.3
colorama~=0.4.6
Jinja2~=3.1.6
numpy~=2.2.4
requests~=2.32.3
djangorestframework~=3.16.0
pytz~=2025.2
six~=1.17.0
pillow~=11.2.1
websocket-client~=1.8.0
channels==4.2.2
python-decouple~=3.8
dj-config-url~=0.1.1
gunicorn
whitenoise
psycopg2-binary